+++ client/src/components/application/Applications.js
@@ -31,53 +31,46 @@ export default function Orders() {
       }
     }
 
+    if (user) {
+      fetchApplications()
+    }
 
-        if (user) {
-            fetchApplications()
-        }
-    }, [dispatch2, user])
+  }, [dispatch, user])
 
-    return (
-        <React.Fragment>
-            <Title>Your Applications</Title>
-            <Table size="small">
-                <TableHead>
-                    <TableRow>
-                        {/* <TableCell>Project ID</TableCell> */}
-                        <TableCell>Title</TableCell>
-                        <TableCell>Offered By</TableCell>
-                        <TableCell>Statement of Purpose</TableCell>
-                        <TableCell>Type</TableCell>
-                        <TableCell>Status</TableCell>
-                    </TableRow>
-                </TableHead>
-                <TableBody>
-                    {applications &&
-                        applications.map(
-                            (application) =>
-                                application.studentEmail === user.email && (
-                                    <TableRow key={application._id}>
-                                        {/* <TableCell>{application.projectID}</TableCell> */}
-                                        <TableCell>{application.projectTitle}</TableCell>
-                                        <TableCell>{application.profEmail}</TableCell>
-                                        <TableCell>{application.sop}</TableCell>
-                                        <TableCell>{application.type == 1 ? "Formal" : "Informal"}</TableCell>
-                                        {/* 0-> undetermined; 1-> accepted, 2-> rejected, 3-> hod approval */}
-                                        <TableCell>
-                                            {application.status == 0
-                                                ? "Applied"
-                                                : application.status === 1
-                                                ? "Accepted"
-                                                : application.status === 2
-                                                ? "Rejected"
-                                                : "Sent for HOD Approval"}
-                                        </TableCell>
-                                    </TableRow>
-                                )
-                        )}
-                </TableBody>
-            </Table>
-            {/* <Link color="primary" href="#" onClick={preventDefault} sx={{ mt: 3 }}>
+  return (
+    <React.Fragment>
+      <Title>Your Applications</Title>
+      <Table size="small">
+        <TableHead>
+          <TableRow>
+            {/* <TableCell>Project ID</TableCell> */}
+            <TableCell>Title</TableCell>
+            <TableCell>Offered By</TableCell>
+            <TableCell>Statement of Purpose</TableCell>
+            <TableCell>Type</TableCell>
+            <TableCell>Status</TableCell>
+          </TableRow>
+        </TableHead>
+        <TableBody>
+          {applications && applications.map((application) => (
+            application.studentEmail === user.email &&
+            <TableRow key={application._id}>
+              {/* <TableCell>{application.projectID}</TableCell> */}
+              <TableCell>{application.projectTitle}</TableCell>
+              <TableCell>{application.profEmail}</TableCell>
+              <TableCell>{application.sop}</TableCell>
+              <TableCell>{application.type == 1 ? "Formal" : "Informal"}</TableCell>
+              {/* 0-> undetermined; 1-> accepted, 2-> rejected, 3-> hod approval */}
+              <TableCell>
+                {application.status == 0 ? "Applied" :
+                  application.status === 1 ? "Accepted" :
+                    application.status === 2 ? "Rejected" : "Sent for HOD Approval"}
+              </TableCell>
+            </TableRow>
+          ))}
+        </TableBody>
+      </Table>
+      {/* <Link color="primary" href="#" onClick={preventDefault} sx={{ mt: 3 }}>
         See more
       </Link> */}
     </React.Fragment>
