+++ client/src/pages/professor/ProfDashboard.js
@@ -34,8 +34,6 @@ function DashboardContent() {
     const [open, setOpen] = React.useState(true)
     const [selectedContent, setSelectedContent] = useState("dashboard")
     const [projectID, setProjectID] = useState(null)
-    const [numberOfStudents, setNumberOfStudents] = useState(null)
-
 
     const handleViewApplicationClick = (content) => {
         setSelectedContent("application")
@@ -143,15 +141,7 @@ function DashboardContent() {
                         {selectedContent === "dashboard" && (
                             <Projects onViewApplicationClick={handleViewApplicationClick} />
                         )}
-
-                        {selectedContent === "application" && (
-                            <ViewApplications
-                                projectID={projectID}
-                                numberOfStudents={numberOfStudents}
-                                onListItemClick={handleListItemClick}
-                            />
-                        )}
-
+                        {selectedContent === "application" && <ViewApplications projectID={projectID} />}
                     </TableContainer>
                 </Box>
             </Box>
