+++ server/server.js
@@ -53,7 +53,7 @@ app.post("/authenticate", async (req, res) => {
         const r = await axios.put(
             "https://api.chatengine.io/users/",
             { username: username, secret: username, first_name: username },
-            { headers: { "Private-Key": "bd81780d-7af4-4add-be57-7bcdd870fbb5" } }
+            { headers: { "Private-Key": "66fb706d-7c8c-4c26-8edc-49dd33ea1038" } }
         )
         return res.status(r.status).json(r.data)
     } catch (e) {
